+++ packages/ui/src/components/ui/badge.tsx
@@ -10,12 +10,12 @@ const badgeVariants = cva(
     variants: {
       variant: {
         default:
-          "border-brand bg-brand-primary-alt text-brand-secondary hover:bg-brand-secondary",
+          "border-brand bg-brand-primary text-brand-secondary hover:bg-brand-secondary",
         secondary:
           "border-secondary bg-secondary text-secondary hover:bg-secondary-hover",
         destructive:
-          "border-error bg-error-primary text-error-primary-hover hover:bg-error-secondary",
-        success: "bg-success-primary text-success-primary hover:bg-success-secondary",
+          "border-error bg-error-secondary text-error-primary hover:bg-error-secondary",
+        success: "bg-success-secondary text-success-primary hover:bg-success-secondary",
         warning: "border-warning bg-warning-secondary text-warning-primary",
         outline: "text-secondary",
       },