+++ packages/ui/src/components/ui/slider.tsx
@@ -18,9 +18,9 @@ const Slider = React.forwardRef<
     {...props}
   >
     <SliderPrimitive.Track className="relative h-1.5 w-full grow overflow-hidden rounded-full bg-quaternary">
-      <SliderPrimitive.Range className="absolute h-full bg-brand" />
+      <SliderPrimitive.Range className="absolute h-full bg-brand-quinary" />
     </SliderPrimitive.Track>
-    <SliderPrimitive.Thumb className="block h-4 w-4 rounded-full border border-brand-alt bg-primary shadow transition-colors focus-visible:outline-none dark:ring-1 dark:ring-ring focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50" />
+    <SliderPrimitive.Thumb className="block h-4 w-4 rounded-full border border-brand bg-primary shadow transition-colors focus-visible:outline-none dark:ring-1 dark:ring-ring focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50" />
   </SliderPrimitive.Root>
 ))
 Slider.displayName = SliderPrimitive.Root.displayName